+++ packages/layout/src/navigation/navigation-section/NavigationSection.tsx
@@ -8,9 +8,7 @@ import { NavigationItem } from '../navigation-item'
 import { SidebarContext } from '../../sidebar'
 import styles from './NavigationSection.module.css'
 
-export interface NavigationSectionProps<
-  T extends object,
-> extends CollectionProps<T> {
+export type NavigationSectionProps<T extends object> = CollectionProps<T> & {
   title?: string
   className?: string
 }