+++ packages/table-styles/CHANGELOG.md
@@ -1,3 +1,27 @@
+## 2.0.105 (2026-08-18)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.21.1
+
+## 2.0.104 (2026-08-18)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.21.0
+
+## 2.0.103 (2026-08-14)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.20.0
+
+## 2.0.102 (2026-08-13)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.19.0
+
 ## 2.0.101 (2026-08-13)
 
 ### 🧱 Updated Dependencies